Added first code
Added basic code to where user can enter a directory path and the names of each program they  want to open/run (programs from other directories can be also be opened by entering the full path of the program).
Added class CCArray which is used to dynamically store an array of strings.